+++ objects/benches/account_seed.rs
@@ -1,6 +1,6 @@
 use criterion::{criterion_group, criterion_main, Criterion};
 use miden_objects::{
-    accounts::{AccountId, AccountStorageMode, AccountType},
+    accounts::{AccountId, AccountStorageMode, AccountType, AccountVersion},
     Digest,
 };
 
@@ -16,6 +16,8 @@ fn grind_account_seed(c: &mut Criterion) {
                 init_seed,
                 AccountType::RegularAccountImmutableCode,
                 AccountStorageMode::Public,
+                AccountVersion::VERSION_0,
+                Digest::default(),
                 Digest::default(),
                 Digest::default(),
             )
@@ -28,6 +30,8 @@ fn grind_account_seed(c: &mut Criterion) {
                 init_seed,
                 AccountType::FungibleFaucet,
                 AccountStorageMode::Public,
+                AccountVersion::VERSION_0,
+                Digest::default(),
                 Digest::default(),
                 Digest::default(),
             )