+++ objects/src/accounts/builder/mod.rs
@@ -6,12 +6,13 @@ use vm_processor::Digest;
 use crate::{
     accounts::{
         Account, AccountCode, AccountComponent, AccountId, AccountStorage, AccountStorageMode,
-        AccountType,
+        AccountType, AccountVersion,
     },
-    assets::{Asset, AssetVault},
-    AccountError, Felt, Word,
+    assets::AssetVault,
+    AccountError, BlockHeader, Felt, Word,
 };
 
+// TODO: Update documentation for newly added fields.
 /// A convenient builder for an [`Account`] allowing for safe construction of an account by
 /// combining multiple [`AccountComponent`]s.
 ///
@@ -36,11 +37,14 @@ use crate::{
 #[derive(Debug, Clone)]
 pub struct AccountBuilder {
     #[cfg(any(feature = "testing", test))]
-    assets: Vec<Asset>,
+    assets: Vec<crate::assets::Asset>,
     components: Vec<AccountComponent>,
     account_type: AccountType,
     storage_mode: AccountStorageMode,
+    block_hash: Digest,
     init_seed: Option<[u8; 32]>,
+    version: AccountVersion,
+    block_epoch: Option<u16>,
 }
 
 impl AccountBuilder {
@@ -51,8 +55,11 @@ impl AccountBuilder {
             assets: vec![],
             components: vec![],
             init_seed: None,
+            block_hash: Digest::default(),
             account_type: AccountType::RegularAccountUpdatableCode,
             storage_mode: AccountStorageMode::Private,
+            version: AccountVersion::VERSION_0,
+            block_epoch: None,
         }
     }
 
@@ -62,6 +69,36 @@ impl AccountBuilder {
     ///  This method **must** be called.
     pub fn init_seed(mut self, init_seed: [u8; 32]) -> Self {
         self.init_seed = Some(init_seed);
+        self
+    }
+
+    /// Sets `block_hash` and `epoch` from the given `block`.
+    ///
+    /// These values must match to create a valid [`AccountId`], so this method is preferred over
+    /// setting the values individually.
+    pub fn with_reference_block(mut self, block: &BlockHeader) -> Self {
+        let block_hash = block.hash();
+        let block_epoch = block.block_epoch();
+        self.block_hash = block_hash;
+        self.block_epoch = Some(block_epoch);
+        self
+    }
+
+    /// Sets the `block_hash` which is an input to the [`AccountId`] derivation process.
+    pub fn block_hash(mut self, block_hash: Digest) -> Self {
+        self.block_hash = block_hash;
+        self
+    }
+
+    /// Sets the [`AccountVersion`] of the account.
+    pub fn version(mut self, version: AccountVersion) -> Self {
+        self.version = version;
+        self
+    }
+
+    /// Sets the `epoch` of the account. Must be less than [`u16::MAX`] else building will fail.
+    pub fn block_epoch(mut self, epoch: u16) -> Self {
+        self.block_epoch = Some(epoch);
         self
     }
 
@@ -120,24 +157,25 @@ impl AccountBuilder {
     fn grind_account_id(
         &self,
         init_seed: [u8; 32],
+        version: AccountVersion,
         code_commitment: Digest,
         storage_commitment: Digest,
-    ) -> Result<(AccountId, Word), AccountError> {
+        block_hash: Digest,
+    ) -> Result<Word, AccountError> {
         let seed = AccountId::get_account_seed(
             init_seed,
             self.account_type,
             self.storage_mode,
+            version,
             code_commitment,
             storage_commitment,
+            block_hash,
         )
         .map_err(|err| {
             AccountError::BuildError("account seed generation failed".into(), Some(Box::new(err)))
         })?;
 
-        let account_id = AccountId::new(seed, code_commitment, storage_commitment)
-            .expect("get_account_seed should provide a suitable seed");
-
-        Ok((account_id, seed))
+        Ok(seed)
     }
 
     /// Builds an [`Account`] out of the configured builder.
@@ -158,6 +196,21 @@ impl AccountBuilder {
     pub fn build(self) -> Result<(Account, Word), AccountError> {
         let (init_seed, vault, code, storage) = self.build_inner()?;
 
+        // Block hash and block epoch must only be set when building a new account.
+        if self.block_hash == Digest::default() {
+            return Err(AccountError::BuildError(
+                "block hash must be set to a `Digest` different from the empty value".into(),
+                None,
+            ));
+        }
+
+        let block_epoch = match self.block_epoch {
+            Some(block_epoch) => block_epoch,
+            None => {
+                return Err(AccountError::BuildError("block epoch must be set".into(), None));
+            },
+        };
+
         #[cfg(any(feature = "testing", test))]
         if !vault.is_empty() {
             return Err(AccountError::BuildError(
@@ -166,8 +219,22 @@ impl AccountBuilder {
             ));
         }
 
-        let (account_id, seed) =
-            self.grind_account_id(init_seed, code.commitment(), storage.commitment())?;
+        let seed = self.grind_account_id(
+            init_seed,
+            self.version,
+            code.commitment(),
+            storage.commitment(),
+            self.block_hash,
+        )?;
+
+        let account_id = AccountId::new(
+            seed,
+            block_epoch,
+            code.commitment(),
+            storage.commitment(),
+            self.block_hash,
+        )
+        .expect("get_account_seed should provide a suitable seed");
 
         debug_assert_eq!(account_id.account_type(), self.account_type);
         debug_assert_eq!(account_id.storage_mode(), self.storage_mode);
@@ -184,7 +251,7 @@ impl AccountBuilder {
     ///
     /// Must only be used when using [`Self::build_existing`] instead of [`Self::build`] since new
     /// accounts must have an empty vault.
-    pub fn with_assets<I: IntoIterator<Item = Asset>>(mut self, assets: I) -> Self {
+    pub fn with_assets<I: IntoIterator<Item = crate::assets::Asset>>(mut self, assets: I) -> Self {
         self.assets.extend(assets);
         self
     }
@@ -198,8 +265,8 @@ impl AccountBuilder {
         let (init_seed, vault, code, storage) = self.build_inner()?;
 
         let account_id = {
-            let bytes = <[u8; 8]>::try_from(&init_seed[0..8])
-                .expect("we should have sliced exactly 8 bytes off");
+            let bytes = <[u8; 15]>::try_from(&init_seed[0..15])
+                .expect("we should have sliced exactly 15 bytes off");
             AccountId::new_with_type_and_mode(bytes, self.account_type, self.storage_mode)
         };
 
@@ -221,6 +288,7 @@ mod tests {
     use std::sync::LazyLock;
 
     use assembly::{Assembler, Library};
+    use assert_matches::assert_matches;
     use vm_core::FieldElement;
 
     use super::*;
@@ -288,8 +356,13 @@ mod tests {
         let storage_slot1 = 12;
         let storage_slot2 = 42;
 
+        let block_hash = Digest::new([Felt::new(42); 4]);
+        let block_epoch = 50_000;
+
         let (account, seed) = Account::builder()
             .init_seed([5; 32])
+            .block_hash(block_hash)
+            .block_epoch(block_epoch)
             .with_component(CustomComponent1 { slot0: storage_slot0 })
             .with_component(CustomComponent2 {
                 slot0: storage_slot1,
@@ -301,8 +374,14 @@ mod tests {
         // Account should be new, i.e. nonce = zero.
         assert_eq!(account.nonce(), Felt::ZERO);
 
-        let computed_id =
-            AccountId::new(seed, account.code.commitment(), account.storage.commitment()).unwrap();
+        let computed_id = AccountId::new(
+            seed,
+            block_epoch,
+            account.code.commitment(),
+            account.storage.commitment(),
+            block_hash,
+        )
+        .unwrap();
         assert_eq!(account.id(), computed_id);
 
         // The merged code should have one procedure from each library.
@@ -353,13 +432,13 @@ mod tests {
 
         let build_error = Account::builder()
             .init_seed([0xff; 32])
+            .block_hash([10; 32].try_into().unwrap())
+            .block_epoch(0)
             .with_component(CustomComponent1 { slot0: storage_slot0 })
             .with_assets(AssetVault::mock().assets())
             .build()
             .unwrap_err();
 
-        assert!(
-            matches!(build_error, AccountError::BuildError(msg, _) if msg == "account asset vault must be empty on new accounts")
-        )
+        assert_matches!(build_error, AccountError::BuildError(msg, _) if msg == "account asset vault must be empty on new accounts")
     }
 }