+++ objects/src/assets/nonfungible.rs
@@ -3,8 +3,9 @@ use core::fmt;
 
 use vm_core::{FieldElement, WORD_SIZE};
 
-use super::{AccountId, AccountType, Asset, AssetError, Felt, Hasher, Word, ACCOUNT_ISFAUCET_MASK};
+use super::{AccountIdPrefix, AccountType, Asset, AssetError, Felt, Hasher, Word};
 use crate::{
+    accounts::AccountId,
     utils::{ByteReader, ByteWriter, Deserializable, DeserializationError, Serializable},
     Digest,
 };
@@ -20,7 +21,7 @@ const FAUCET_ID_POS: usize = 1;
 ///
 /// - Hash the asset data producing `[d0, d1, d2, d3]`.
 /// - Replace the value of `d1` with the faucet id producing `[d0, faucet_id, d2, d3]`.
-/// - Force the bit position [ACCOUNT_ISFAUCET_MASK] of `d3` to be `0`.
+/// - Force the bit position [`AccountId::IS_FAUCET_MASK`] of `d3` to be `0`.
 ///
 /// [NonFungibleAsset] itself does not contain the actual asset data. The container for this data
 /// [NonFungibleAssetDetails] struct.
@@ -68,13 +69,13 @@ impl NonFungibleAsset {
     ///
     /// # Errors
     /// Returns an error if the provided faucet ID is not for a non-fungible asset faucet.
-    pub fn from_parts(faucet_id: AccountId, mut data_hash: Word) -> Result<Self, AssetError> {
+    pub fn from_parts(faucet_id: AccountIdPrefix, mut data_hash: Word) -> Result<Self, AssetError> {
         if !matches!(faucet_id.account_type(), AccountType::NonFungibleFaucet) {
             return Err(AssetError::NonFungibleFaucetIdTypeMismatch(faucet_id));
         }
         data_hash[FAUCET_ID_POS] = faucet_id.into();
 
-        // Forces the bit at position `ACCOUNT_ISFAUCET_MASK` to `0`.
+        // Forces the bit at position `AccountId::IS_FAUCET_MASK` to `0`.
         //
         // Explanation of the bit flip:
         //
@@ -91,8 +92,10 @@ impl NonFungibleAsset {
         // merging of fungible faucets. At the same time, it spreads the non-fungible assets evenly
         // across the vault, because in this case the element is the result of a cryptographic hash
         // function.
-        let d3 = data_hash[3].as_int();
-        data_hash[3] = Felt::new((d3 & ACCOUNT_ISFAUCET_MASK) ^ d3);
+
+        // TODO: Update explanation above.
+        let element3 = data_hash[3].as_int();
+        data_hash[3] = Felt::new((element3 & AccountId::IS_FAUCET_MASK) ^ element3);
 
         let asset = Self(data_hash);
 
@@ -115,8 +118,8 @@ impl NonFungibleAsset {
     }
 
     /// Return ID of the faucet which issued this asset.
-    pub fn faucet_id(&self) -> AccountId {
-        AccountId::new_unchecked(self.0[FAUCET_ID_POS])
+    pub fn faucet_id(&self) -> AccountIdPrefix {
+        AccountIdPrefix::new_unchecked(self.0[FAUCET_ID_POS])
     }
 
     // HELPER FUNCTIONS
@@ -128,7 +131,7 @@ impl NonFungibleAsset {
     /// - The faucet_id is not a valid non-fungible faucet ID.
     /// - The most significant bit of the asset is not ZERO.
     fn validate(&self) -> Result<(), AssetError> {
-        let faucet_id = AccountId::try_from(self.0[FAUCET_ID_POS])
+        let faucet_id = AccountIdPrefix::try_from(self.0[FAUCET_ID_POS])
             .map_err(|err| AssetError::InvalidFaucetAccountId(Box::new(err)))?;
 
         let account_type = faucet_id.account_type();
@@ -188,27 +191,15 @@ impl Serializable for NonFungibleAsset {
 
 impl Deserializable for NonFungibleAsset {
     fn read_from<R: ByteReader>(source: &mut R) -> Result<Self, DeserializationError> {
-        let faucet_id: AccountId = source.read()?;
+        let faucet_id_prefix: AccountIdPrefix = source.read()?;
 
-        Self::deserialize_with_account_id(faucet_id, source)
-            .map_err(|err| DeserializationError::InvalidValue(err.to_string()))
-    }
-}
-
-impl NonFungibleAsset {
-    /// Deserializes a [`NonFungibleAsset`] from an [`AccountId`] and the remaining data from the
-    /// given `source`.
-    pub(super) fn deserialize_with_account_id<R: ByteReader>(
-        faucet_id: AccountId,
-        source: &mut R,
-    ) -> Result<Self, DeserializationError> {
         let hash_0: Felt = source.read()?;
         let hash_2: Felt = source.read()?;
         let hash_3: Felt = source.read()?;
 
         // The second felt in the data_hash will be replaced by the faucet id, so we can set it to
         // zero here.
-        NonFungibleAsset::from_parts(faucet_id, [hash_0, Felt::ZERO, hash_2, hash_3])
+        NonFungibleAsset::from_parts(faucet_id_prefix, [hash_0, Felt::ZERO, hash_2, hash_3])
             .map_err(|err| DeserializationError::InvalidValue(err.to_string()))
     }
 }
@@ -221,7 +212,7 @@ impl NonFungibleAsset {
 /// Unlike [NonFungibleAsset] struct, this struct contains full details of a non-fungible asset.
 #[derive(Debug, Clone, PartialEq, Eq)]
 pub struct NonFungibleAssetDetails {
-    faucet_id: AccountId,
+    faucet_id: AccountIdPrefix,
     asset_data: Vec<u8>,
 }
 
@@ -230,7 +221,7 @@ impl NonFungibleAssetDetails {
     ///
     /// # Errors
     /// Returns an error if the provided faucet ID is not for a non-fungible asset faucet.
-    pub fn new(faucet_id: AccountId, asset_data: Vec<u8>) -> Result<Self, AssetError> {
+    pub fn new(faucet_id: AccountIdPrefix, asset_data: Vec<u8>) -> Result<Self, AssetError> {
         if !matches!(faucet_id.account_type(), AccountType::NonFungibleFaucet) {
             return Err(AssetError::NonFungibleFaucetIdTypeMismatch(faucet_id));
         }
@@ -239,7 +230,7 @@ impl NonFungibleAssetDetails {
     }
 
     /// Returns ID of the faucet which issued this asset.
-    pub fn faucet_id(&self) -> AccountId {
+    pub fn faucet_id(&self) -> AccountIdPrefix {
         self.faucet_id
     }
 
@@ -254,10 +245,15 @@ impl NonFungibleAssetDetails {
 
 #[cfg(test)]
 mod tests {
+    use assert_matches::assert_matches;
+
     use super::*;
-    use crate::accounts::account_id::testing::{
-        ACCOUNT_ID_FUNGIBLE_FAUCET_OFF_CHAIN, ACCOUNT_ID_NON_FUNGIBLE_FAUCET_OFF_CHAIN,
-        ACCOUNT_ID_NON_FUNGIBLE_FAUCET_ON_CHAIN, ACCOUNT_ID_NON_FUNGIBLE_FAUCET_ON_CHAIN_1,
+    use crate::accounts::{
+        account_id::testing::{
+            ACCOUNT_ID_FUNGIBLE_FAUCET_OFF_CHAIN, ACCOUNT_ID_NON_FUNGIBLE_FAUCET_OFF_CHAIN,
+            ACCOUNT_ID_NON_FUNGIBLE_FAUCET_ON_CHAIN, ACCOUNT_ID_NON_FUNGIBLE_FAUCET_ON_CHAIN_1,
+        },
+        AccountId,
     };
 
     #[test]
@@ -268,7 +264,7 @@ mod tests {
             ACCOUNT_ID_NON_FUNGIBLE_FAUCET_ON_CHAIN_1,
         ] {
             let account_id = AccountId::try_from(non_fungible_account_id).unwrap();
-            let details = NonFungibleAssetDetails::new(account_id, vec![1, 2, 3]).unwrap();
+            let details = NonFungibleAssetDetails::new(account_id.prefix(), vec![1, 2, 3]).unwrap();
             let non_fungible_asset = NonFungibleAsset::new(&details).unwrap();
             assert_eq!(
                 non_fungible_asset,
@@ -277,12 +273,16 @@ mod tests {
         }
 
         let account = AccountId::try_from(ACCOUNT_ID_NON_FUNGIBLE_FAUCET_OFF_CHAIN).unwrap();
-        let details = NonFungibleAssetDetails::new(account, vec![4, 5, 6, 7]).unwrap();
+        let details = NonFungibleAssetDetails::new(account.prefix(), vec![4, 5, 6, 7]).unwrap();
         let asset = NonFungibleAsset::new(&details).unwrap();
         let mut asset_bytes = asset.to_bytes();
-        // Set invalid Faucet ID.
-        asset_bytes[0..8].copy_from_slice(&ACCOUNT_ID_FUNGIBLE_FAUCET_OFF_CHAIN.to_le_bytes());
+
+        let fungible_faucet_id = AccountId::try_from(ACCOUNT_ID_FUNGIBLE_FAUCET_OFF_CHAIN).unwrap();
+
+        // Set invalid Faucet ID Prefix.
+        asset_bytes[0..8].copy_from_slice(&fungible_faucet_id.prefix().to_bytes());
+
         let err = NonFungibleAsset::read_from_bytes(&asset_bytes).unwrap_err();
-        assert!(matches!(err, DeserializationError::InvalidValue(_)));
+        assert_matches!(err, DeserializationError::InvalidValue(msg) if msg.contains("must be of type NonFungibleFaucet"));
     }
 }