+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 - Added conversion from `Account` to `AccountDelta` for initial account state representation as delta (#983).
 - [BREAKING] Added `miden::note::get_script_hash` procedure (#995).
 - [BREAKING] Refactor error messages in `miden-lib` and `miden-tx` and use `thiserror` 2.0 (#1005).
+- [BREAKING] Extend `AccountId` to two `Felt`s and require block hash in derivation (#982).
 
 ## 0.6.2 (2024-11-20)